# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: SafeSchool safeguarding and incident reporting API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/pupil/login:
    post:
      operationId: pupilLogin
      tags: [auth]
      summary: Pupil login with PIN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PupilLoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/staff/login:
    post:
      operationId: staffLogin
      tags: [auth]
      summary: Staff login with email/password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffLoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/parent/login:
    post:
      operationId: parentLogin
      tags: [auth]
      summary: Parent login with email/password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffLoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /alerts:
    get:
      operationId: listAlerts
      tags: [alerts]
      summary: List pattern alerts
      parameters:
        - name: level
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated alerts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedAlerts"
  /alerts/{id}:
    patch:
      operationId: updateAlert
      tags: [alerts]
      summary: Update alert status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAlertBody"
      responses:
        "200":
          description: Alert updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatternAlert"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    PupilLoginBody:
      type: object
      properties:
        schoolId:
          type: string
        pupilId:
          type: string
        pin:
          type: string
      required:
        - schoolId
        - pupilId
        - pin
    StaffLoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
        firstLogin:
          type: boolean
      required:
        - token
        - user
        - firstLogin
    User:
      type: object
      properties:
        id:
          type: string
        schoolId:
          type: string
        role:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: ["string", "null"]
        yearGroup:
          type: ["string", "null"]
        className:
          type: ["string", "null"]
        avatarType:
          type: ["string", "null"]
        avatarValue:
          type: ["string", "null"]
        avatarImageUrl:
          type: ["string", "null"]
        parentOf:
          type: array
          items:
            type: string
        active:
          type: boolean
        lastLogin:
          type: ["string", "null"]
      required:
        - id
        - schoolId
        - role
        - firstName
        - lastName
        - active
...
    PatternAlert:
      type: object
      properties:
        id:
          type: string
        schoolId:
          type: string
        ruleId:
          type: string
        ruleLabel:
          type: ["string", "null"]
        alertLevel:
          type: string
        victimId:
          type: ["string", "null"]
        perpetratorIds:
          type: array
          items:
            type: string
        linkedIncidentIds:
          type: array
          items:
            type: string
        triggeredAt:
          type: string
        reviewedAt:
          type: ["string", "null"]
        reviewedBy:
          type: ["string", "null"]
        status:
          type: string
        notes:
          type: ["string", "null"]
        victimName:
          type: ["string", "null"]
      required:
        - id
        - schoolId
        - ruleId
        - alertLevel
        - triggeredAt
        - status
    PaginatedAlerts:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/PatternAlert"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit
    UpdateAlertBody:
      type: object
      properties:
        status:
          type: string
        notes:
          type: ["string", "null"]
      required:
        - status
...